server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.5
aiofiles>=0.8.0
requests>=2.28.0
beautifulsoup4>=4.11.0
cloudscraper>=1.2.58
numpy>=1.24.0
Pillow>=9.0.0
nameparser>=1.1.0
websockets>=10.4
scikit-learn>=1.0.0 